extracting class to a file
